fixed notifications

+++ auctions/ajax_controls.py
@@ -39,11 +39,10 @@ def ajax_delete_comment(request):
 def ajax_dismiss_notification(request):
     notification_id = request.POST.get('notification_id', None)
     if notification_id:
-        notification = get_object_or_404(Notification, pk=notification_id)
-        notification.delete()
-        return JsonResponse({}, status=200)
-    else:
-        return JsonResponse({}, status=404)
+        notification = Notification.objects.get(pk=notification_id)
+        if notification:
+            notification.delete()
+    return JsonResponse({}, status=200)
 
 
 def ajax_generate_comment(request):
